requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
idna==3.4
importlib-metadata==4.8.3
pika==1.3.1
#pkg-resources==0.0.0
prettytable==2.5.0
# psycopg2==2.9.7
psycopg2-binary #in Github Actions Pipeline
PyJWT==2.4.0
pytz==2023.3
requests==2.27.1
twilio==7.17.0
typing-extensions==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.6.0
